requirements.txt
Django==3.0.4
djangorestframework==3.11.0
django-environ==0.4.5
twilio==6.37.0
